Working mongoDB local server. With visualizations
