+++ frontend/src/components/container/App.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import {any } from "prop-types";
 import NavigationContainer from "./NavigationContainer";
-import DashboardGrid from "./DashboardContainer";
 import '../App.scss';
 
 export default class App extends Component {
